requirements.txt
# general modules
Django~=4.2.17
python-decouple~=3.8
Pillow

# third party modules
django-debug-toolbar~=3.2.1
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter
drf-yasg[validation]
djoser

# deployment module


# email third-party modules
django-mail-templated

# refactoring and tests
black